+++ packages/tiny-rpc/src/adapter-web.test.ts
@@ -1,84 +1,10 @@
-import { AsyncLocalStorage } from "node:async_hooks";
 import { createServer } from "@hattip/adapter-node";
-import { type RequestHandler, compose } from "@hattip/compose";
+import { compose } from "@hattip/compose";
 import { tinyassert } from "@hiogawa/utils";
 import { describe, expect, it, vi } from "vitest";
-import { z } from "zod";
 import { httpClientAdapter, httpServerAdapter } from "./adapter-web";
-import {
-  TinyRpcError,
-  type TinyRpcRoutes,
-  exposeTinyRpc,
-  proxyTinyRpc,
-} from "./core";
-import { validateFn } from "./validation";
-
-//
-// example rpc
-//
-
-function defineExampleRpc() {
-  //
-  // service internal
-  //
-  let counter = 0;
-
-  //
-  // context
-  // (implicit input e.g. request session etc...)
-  // (it can be scoped by request using AsyncLocalStorage and middleware)
-  //
-  interface RpcContext {
-    request: Request;
-  }
-
-  const contextStorage = new AsyncLocalStorage<RpcContext>();
-
-  function useContext() {
-    const ctx = contextStorage.getStore();
-    tinyassert(ctx);
-    return ctx;
-  }
-
-  function contextProviderHandler(): RequestHandler {
-    // middleware to initialize context scoped by request
-    return async (ctx) => {
-      return contextStorage.run({ request: ctx.request }, () => ctx.next());
-    };
-  }
-
-  //
-  // service api
-  //
-  const routes = {
-    // define as a bare function
-    checkId: (id: string) => id === "good",
-
-    checkIdThrow: (id: string) => {
-      tinyassert(id === "good", "Invalid ID");
-      return null;
-    },
-
-    getCounter: () => counter,
-
-    // define with zod validation + input type inference
-    incrementCounter: validateFn(z.object({ delta: z.number().default(1) }))(
-      (input) => {
-        input satisfies { delta: number };
-        counter += input.delta;
-        return counter;
-      }
-    ),
-
-    // access context
-    checkAuth: () => {
-      const { request } = useContext();
-      return request.headers.get("x-auth") === "good";
-    },
-  } satisfies TinyRpcRoutes;
-
-  return { routes, contextProviderHandler };
-}
+import { TinyRpcError, exposeTinyRpc, proxyTinyRpc } from "./core";
+import { defineTestRpcRoutes } from "./tests/helper";
 
 //
 // test
@@ -86,9 +12,8 @@ function defineExampleRpc() {
 
 describe("e2e", () => {
   it("basic", async () => {
-    // define example rpc
     const endpoint = "/rpc";
-    const { routes, contextProviderHandler } = defineExampleRpc();
+    const { routes, contextProviderHandler } = defineTestRpcRoutes();
 
     //
     // server
@@ -212,9 +137,8 @@ describe("e2e", () => {
 
   // test is copy-pasted for GET
   it("GET", async () => {
-    // define example rpc
     const endpoint = "/rpc";
-    const { routes, contextProviderHandler } = defineExampleRpc();
+    const { routes, contextProviderHandler } = defineTestRpcRoutes();
 
     //
     // server
